requirements.txt
# AgriDec - 农业数据服务平台核心依赖包

# Web框架
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Flask-Bcrypt==1.0.1

# 数据库支持
PyMySQL==1.1.0
SQLAlchemy==2.0.21

# 环境配置
python-dotenv==1.0.0

# 数据处理和分析
pandas==2.1.1
numpy==1.24.3

# 网络请求和爬虫
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# 日期时间处理
python-dateutil==2.8.2

# 定时任务调度
APScheduler==3.10.4

# Web安全
Werkzeug==2.3.7

# 模板引擎
Jinja2==3.1.2
MarkupSafe==2.1.3

# 其他核心工具
click==8.1.7
itsdangerous==2.1.2

# 测试框架
pytest==7.4.2
pytest-flask==1.2.0
playwright==1.40.0

# 生产部署
gunicorn==21.2.0

# 安全
cryptography==41.0.4
